Refactor erc20, erc721 transfers
