+++ src/pages/AddressPage/tabs/Transactions.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from "react";
-import { Box, ColumnConfig, Text, Tip } from "grommet";
+import { Box } from "grommet";
 import { useParams } from "react-router-dom";
 import {
   getByteCodeSignatureByHash,
@@ -8,379 +8,16 @@ import {
 } from "src/api/client";
 import { TransactionsTable } from "src/components/tables/TransactionsTable";
 import {
-  Address,
-  ONEValue,
-  DateTime,
-} from "src/components/ui";
-import {
   Filter,
   RelatedTransaction,
   RelatedTransactionType, RPCTransactionHarmony
 } from "src/types";
-import styled, { css } from "styled-components";
 import { TRelatedTransaction } from "src/api/client.interface";
-import { getERC20Columns } from "./erc20Columns";
 import { getAddress } from "src/utils";
 import { ExportToCsvButton } from "../../../components/ui/ExportToCsvButton";
-
-const TxMethod = styled(Text)`
-  width: 100px;
-
-  > div {
-    display: block;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-`
-
-const Marker = styled.div<{ out: boolean }>`
-  border-radius: 2px;
-  padding: 5px;
-
-  text-align: center;
-  font-weight: bold;
-
-  ${(props) =>
-    props.out
-      ? css`
-          background: rgb(239 145 62);
-          color: #fff;
-        `
-      : css`
-          background: rgba(105, 250, 189, 0.8);
-          color: #1b295e;
-        `};
-`;
-
-const NeutralMarker = styled(Box)`
-  border-radius: 2px;
-  padding: 5px;
-
-  text-align: center;
-  font-weight: bold;
-`;
+import { getColumns, getERC20Columns, getNFTColumns, getStackingColumns } from "./txsColumns";
 
 const internalTxsBlocksFrom = 23000000
-
-function getColumns(id: string): ColumnConfig<any>[] {
-  return [
-    // {
-    //   property: "type",
-    //   size: "",
-    //   header: (
-    //     <Text
-    //       color="minorText"
-    //       size="small"
-    //       style={{ fontWeight: 300, width: "140px" }}
-    //     >
-    //       Type
-    //     </Text>
-    //   ),
-    //   render: (data: RelatedTransaction) => (
-    //     <Text size="small" style={{ width: "140px" }}>
-    //       {relatedTxMap[data.transactionType] || data.transactionType}
-    //     </Text>
-    //   ),
-    // },
-    {
-      property: "hash",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "95px" }}
-        >
-          Hash
-        </Text>
-      ),
-      render: (data: any) => (
-        <Address
-          address={data.transactionHash || data.hash}
-          type="tx"
-          isShortEllipsis={true}
-          style={{ width: "170px" }}
-        />
-      ),
-    },
-    {
-      property: "method",
-      header: (
-        <Text color="minorText" size="small" style={{ fontWeight: 300 }}>
-          Method
-        </Text>
-      ),
-      render: (data: any) => {
-        let signature;
-
-        try {
-          // @ts-ignore
-          signature =
-            data.signatures &&
-            data.signatures.map((s: any) => s.signature)[0].split("(")[0];
-        } catch (err) {}
-
-        if (!signature && data.value !== "0") {
-          signature = "transfer";
-        }
-
-        if (!signature && data.input.length >= 10) {
-          signature = data.input.slice(2, 10);
-        }
-
-        if (!signature) {
-          return <Text size="small">{"—"}</Text>;
-        }
-
-        return (
-          <Tip content={<span>{signature}</span>}>
-            <TxMethod size="10px">
-              <NeutralMarker background={"backgroundBack"}>
-                {signature}
-              </NeutralMarker>
-            </TxMethod>
-          </Tip>
-        );
-      },
-    },
-    // {
-    //   property: "shard",
-    //   header: (
-    //     <Text color="minorText" size="small" style={{ fontWeight: 300 }}>
-    //       Shard
-    //     </Text>
-    //   ),
-    //   render: (data: RelatedTransaction) => (
-    //     <Box direction="row" gap="3px" align="center">
-    //       <Text size="small">{0}</Text>
-    //       <FormNextLink
-    //         size="small"
-    //         color="brand"
-    //         style={{ marginBottom: "2px" }}
-    //       />
-    //       <Text size="small">{0}</Text>
-    //     </Box>
-    //   ),
-    // },
-    {
-      property: "from",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "180px" }}
-        >
-          From
-        </Text>
-      ),
-      render: (data: RelatedTransaction) => (
-        <Text size="12px">
-          <Address address={data.from} isShortEllipsis={true} style={{ width: '180px' }} />
-        </Text>
-      ),
-    },
-    {
-      property: "marker",
-      header: <></>,
-      render: (data: RelatedTransaction) => (
-        <Text size="12px">
-          <Marker out={data.from === id}>
-            {data.from === id ? "OUT" : "IN"}
-          </Marker>
-        </Text>
-      ),
-    },
-    {
-      property: "to",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "180px" }}
-        >
-          To
-        </Text>
-      ),
-      render: (data: RelatedTransaction) => (
-        <Text size="12px">
-          <Address address={data.to} isShortEllipsis={true} style={{ width: '180px' }} />
-        </Text>
-      ),
-    },
-    {
-      property: "value",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "120px" }}
-        >
-          Value
-        </Text>
-      ),
-      render: (data: RelatedTransaction) => (
-        <Box justify="center">
-          <ONEValue value={data.value} timestamp={data.timestamp} />
-        </Box>
-      ),
-    },
-
-    {
-      property: "timestamp",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "140px" }}
-        >
-          Timestamp
-        </Text>
-      ),
-      render: (data: RelatedTransaction) => (
-        <Box direction="row" gap="xsmall" justify="end">
-          <DateTime
-            date={data.timestamp}
-          />
-        </Box>
-      ),
-    },
-  ];
-}
-
-const getStackingColumns = (id: string): ColumnConfig<any>[] => {
-  return [
-    {
-      property: "hash",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "95px" }}
-        >
-          Hash
-        </Text>
-      ),
-      render: (data: any) => (
-        <Address
-          address={data.transactionHash || data.hash}
-          type="staking-tx"
-          isShortEllipsis={true}
-          style={{ width: "170px" }}
-        />
-      ),
-    },
-    {
-      property: "type",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "140px" }}
-        >
-          Type
-        </Text>
-      ),
-      render: (data: RelatedTransaction) => (
-        <Text size="small" style={{ width: "140px" }}>
-          {data.type}
-        </Text>
-      ),
-    },
-    {
-      property: "validator",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "170px" }}
-        >
-          Validator
-        </Text>
-      ),
-      render: (data: RelatedTransaction) => (
-        <Text size="12px">
-          {data.msg?.validatorAddress ? (
-            <Address address={data.msg?.validatorAddress || data.from} isShortEllipsis={true} style={{ width: "170px" }} />
-          ) : (
-            "—"
-          )}
-        </Text>
-      ),
-    },
-    {
-      property: "marker",
-      header: <></>,
-      render: (data: RelatedTransaction) => (
-        <Text size="12px">
-          <Marker out={data.from === id}>
-            {data.from === id ? "OUT" : "IN"}
-          </Marker>
-        </Text>
-      ),
-    },
-    {
-      property: "delegator",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "170px" }}
-        >
-          Delegator
-        </Text>
-      ),
-      render: (data: RelatedTransaction) => (
-        <Text size="12px">
-          {data.msg?.delegatorAddress ? (
-            <Address address={data.msg?.delegatorAddress} isShortEllipsis={true} style={{ width: "170px" }} />
-          ) : (
-            "—"
-          )}
-        </Text>
-      ),
-    },
-    {
-      property: "value",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "220px" }}
-        >
-          Value
-        </Text>
-      ),
-      render: (data: RelatedTransaction) => (
-        <Box justify="center">
-          {data.msg?.amount ? (
-            <ONEValue value={data.msg?.amount} timestamp={data.timestamp} />
-          ) : data.amount ? (
-            <ONEValue value={data.amount} timestamp={data.timestamp} />
-          ) : (
-            "—"
-          )}
-        </Box>
-      ),
-    },
-    {
-      property: "timestamp",
-      header: (
-        <Text
-          color="minorText"
-          size="small"
-          style={{ fontWeight: 300, width: "140px" }}
-        >
-          Timestamp
-        </Text>
-      ),
-      render: (data: RelatedTransaction) => (
-        <Box direction="row" gap="xsmall" justify="end">
-          <DateTime date={data.timestamp} />
-        </Box>
-      ),
-    },
-  ];
-};
 
 const relatedTxMap: Record<RelatedTransactionType, string> = {
   transaction: "Transaction",
@@ -543,6 +180,11 @@ export function Transactions(props: {
     }
     case "erc20": {
       columns = getERC20Columns(id);
+      break;
+    }
+
+    case "erc721": {
+      columns = getNFTColumns(id)
       break;
     }
